requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
python-decouple==3.8
django-filter==23.3
requests==2.31.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
django-import-export==3.3.4
django-unfold==0.20.0
drf-yasg==1.21.7
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
